Add Yaml helper file and tidy serialiser code
