+++ Sprocket/Scene/Serialiser.cpp
@@ -2,199 +2,12 @@
 #include "Log.h"
 #include "Components.h"
 #include "Maths.h"
+#include "Yaml.h"
 
 #include <yaml-cpp/yaml.h>
-
 #include <fstream>
 
-namespace YAML {
-
-template<>
-struct convert<Sprocket::Maths::vec2>
-{
-    static Node encode(const Sprocket::Maths::vec2& rhs)
-    {
-        Node n;
-        n.push_back(rhs.x);
-        n.push_back(rhs.y);
-        return n;
-    }
-
-    static bool decode(const Node& node, Sprocket::Maths::vec2& rhs)
-    {
-        if (!node.IsSequence() || node.size() != 2)
-            return false;
-
-        rhs.x = node[0].as<float>();
-        rhs.y = node[1].as<float>();
-        return true;
-    }
-};
-
-template<>
-struct convert<Sprocket::Maths::vec3>
-{
-    static Node encode(const Sprocket::Maths::vec3& rhs)
-    {
-        Node n;
-        n.push_back(rhs.x);
-        n.push_back(rhs.y);
-        n.push_back(rhs.z);
-        return n;
-    }
-
-    static bool decode(const Node& node, Sprocket::Maths::vec3& rhs)
-    {
-        if (!node.IsSequence() || node.size() != 3)
-            return false;
-
-        rhs.x = node[0].as<float>();
-        rhs.y = node[1].as<float>();
-        rhs.z = node[2].as<float>();
-        return true;
-    }
-};
-
-template<>
-struct convert<Sprocket::Maths::quat>
-{
-    static Node encode(const Sprocket::Maths::quat& rhs)
-    {
-        Node n;
-        n.push_back(rhs.w);
-        n.push_back(rhs.x);
-        n.push_back(rhs.y);
-        n.push_back(rhs.z);
-        return n;
-    }
-
-    static bool decode(const Node& node, Sprocket::Maths::quat& rhs)
-    {
-        if (!node.IsSequence() || node.size() != 4)
-            return false;
-
-        rhs.w = node[0].as<float>();
-        rhs.x = node[1].as<float>();
-        rhs.y = node[2].as<float>();
-        rhs.z = node[3].as<float>();
-        return true;
-    }
-};
-
-template<>
-struct convert<Sprocket::Maths::mat4>
-{
-    static Node encode(const Sprocket::Maths::mat4& rhs)
-    {
-        Node n;
-        n.push_back(rhs[0][0]);
-        n.push_back(rhs[0][1]);
-        n.push_back(rhs[0][2]);
-        n.push_back(rhs[0][3]);
-
-        n.push_back(rhs[1][0]);
-        n.push_back(rhs[1][1]);
-        n.push_back(rhs[1][2]);
-        n.push_back(rhs[1][3]);
-
-        n.push_back(rhs[2][0]);
-        n.push_back(rhs[2][1]);
-        n.push_back(rhs[2][2]);
-        n.push_back(rhs[2][3]);
-
-        n.push_back(rhs[3][0]);
-        n.push_back(rhs[3][1]);
-        n.push_back(rhs[3][2]);
-        n.push_back(rhs[3][3]);
-        return n;
-    }
-
-    static bool decode(const Node& node, Sprocket::Maths::mat4& rhs)
-    {
-        if (!node.IsSequence() || node.size() != 16)
-            return false;
-
-        rhs[0][0] = node[0].as<float>();
-        rhs[0][1] = node[1].as<float>();
-        rhs[0][2] = node[2].as<float>();
-        rhs[0][3] = node[3].as<float>();
-        rhs[1][0] = node[4].as<float>();
-        rhs[1][1] = node[5].as<float>();
-        rhs[1][2] = node[6].as<float>();
-        rhs[1][3] = node[7].as<float>();
-        rhs[2][0] = node[8].as<float>();
-        rhs[2][1] = node[9].as<float>();
-        rhs[2][2] = node[10].as<float>();
-        rhs[2][3] = node[11].as<float>();
-        rhs[3][0] = node[12].as<float>();
-        rhs[3][1] = node[13].as<float>();
-        rhs[3][2] = node[14].as<float>();
-        rhs[3][3] = node[15].as<float>();
-        return true;
-    }
-};
-
-}
-
 namespace Sprocket {
-namespace {
-
-YAML::Emitter& operator<<(YAML::Emitter& out, const Maths::vec3& v)
-{
-    out << YAML::Flow;
-    out << YAML::BeginSeq << v.x << v.y << v.z << YAML::EndSeq;
-    return out;
-}
-
-YAML::Emitter& operator<<(YAML::Emitter& out, const Maths::quat& q)
-{
-    out << YAML::Flow;
-    out << YAML::BeginSeq << q.w << q.x << q.y << q.z << YAML::EndSeq;
-    return out;
-}
-
-YAML::Emitter& operator<<(YAML::Emitter& out, const Maths::mat4& m)
-{
-    out << YAML::Flow;
-    out << YAML::BeginSeq << m[0][0] << m[0][1] << m[0][2] << m[0][3]
-                          << m[1][0] << m[1][1] << m[1][2] << m[1][3]
-                          << m[2][0] << m[2][1] << m[2][2] << m[2][3]
-                          << m[3][0] << m[3][1] << m[3][2] << m[3][3]
-                          << YAML::EndSeq;
-    return out;
-}
-
-template <typename T>
-YAML::Emitter& operator<<(YAML::Emitter& out, const std::queue<T> q)
-{
-    auto copy = q;
-    out << YAML::BeginSeq;
-    while (!copy.empty()) {
-        auto obj = copy.front();
-        copy.pop();
-        out << obj;
-    }
-    out << YAML::EndSeq;
-    return out;
-}
-
-template <typename T>
-struct MapEntry
-{
-    const char* key;
-    T           value;
-
-    MapEntry(const char* key_, const T& value_) : key(key_), value(value_) {}
-};
-
-template <typename T>
-YAML::Emitter& operator<<(YAML::Emitter& out, const MapEntry<T>& entry)
-{
-    out << YAML::Key << entry.key << YAML::Value << entry.value;
-    return out;
-}
-
-}
 
 void Serialiser::Serialise(const std::string& file)
 {
@@ -211,73 +24,67 @@ void Serialiser::Serialise(const std::string& file)
         if (entity.Has<NameComponent>()) {
             const auto& name = entity.Get<NameComponent>();
             out << YAML::Key << "Name" << YAML::BeginMap;
-            out << MapEntry("Name", name.name);
+            out << YAML::Key << "Name" << YAML::Value << name.name;
             out << YAML::EndMap;
         }
         if (entity.Has<TransformComponent>()) {
             const auto& transform = entity.Get<TransformComponent>();
-            out << YAML::Key << "Transform";
-            out << YAML::BeginMap;
-            out << MapEntry("Position", transform.position);
-            out << MapEntry("Orientation", transform.orientation);
+            out << YAML::Key << "Transform" << YAML::BeginMap;
+            out << YAML::Key << "Position" << YAML::Value << transform.position;
+            out << YAML::Key << "Orientation" << YAML::Value << transform.orientation;
             out << YAML::EndMap;
         }
         if (entity.Has<ModelComponent>()) {
             const auto& model = entity.Get<ModelComponent>();
-            out << YAML::Key << "Model";
-            out << YAML::BeginMap;
-            out << MapEntry("Model", model.model);
-            out << MapEntry("Scale", model.scale);
-            out << MapEntry("Texture", model.texture);
-            out << MapEntry("ShineDamper", model.shineDamper);
-            out << MapEntry("Reflectivity", model.reflectivity);
+            out << YAML::Key << "Model" << YAML::BeginMap;
+            out << YAML::Key << "Model" << YAML::Value << model.model;
+            out << YAML::Key << "Scale" << YAML::Value << model.scale;
+            out << YAML::Key << "Texture" << YAML::Value << model.texture;
+            out << YAML::Key << "ShineDamper" << YAML::Value << model.shineDamper;
+            out << YAML::Key << "Reflectivity" << YAML::Value << model.reflectivity;
             out << YAML::EndMap;
         }
         if (entity.Has<PhysicsComponent>()) {
             const auto& physics = entity.Get<PhysicsComponent>();
-            out << YAML::Key << "Physics";
-            out << YAML::BeginMap;
-            out << MapEntry("Velocity", physics.velocity);
-            out << MapEntry("Gravity", physics.gravity);
-            out << MapEntry("Frozen", physics.frozen);
-            out << MapEntry("Collider", static_cast<int>(physics.collider));
-            out << MapEntry("HalfExtents", physics.halfExtents);
-            out << MapEntry("Radius", physics.radius);
-            out << MapEntry("Height", physics.height);
-            out << MapEntry("Mass", physics.mass);
-            out << MapEntry("Bounciness", physics.bounciness);
-            out << MapEntry("FrictionCoefficient", physics.frictionCoefficient);
-            out << MapEntry("RollingResistance", physics.rollingResistance);
+            out << YAML::Key << "Physics" << YAML::BeginMap;
+            out << YAML::Key << "Velocity" << YAML::Value << physics.velocity;
+            out << YAML::Key << "Gravity" << YAML::Value << physics.gravity;
+            out << YAML::Key << "Frozen" << YAML::Value << physics.frozen;
+            out << YAML::Key << "Collider" << YAML::Value << static_cast<int>(physics.collider);
+            out << YAML::Key << "HalfExtents" << YAML::Value << physics.halfExtents;
+            out << YAML::Key << "Radius" << YAML::Value << physics.radius;
+            out << YAML::Key << "Height" << YAML::Value << physics.height;
+            out << YAML::Key << "Mass" << YAML::Value << physics.mass;
+            out << YAML::Key << "Bounciness" << YAML::Value << physics.bounciness;
+            out << YAML::Key << "FrictionCoefficient" << YAML::Value << physics.frictionCoefficient;
+            out << YAML::Key << "RollingResistance" << YAML::Value << physics.rollingResistance;
             out << YAML::EndMap;
         }
         if (entity.Has<ScriptComponent>()) {
             const auto& script = entity.Get<ScriptComponent>();
-            out << YAML::Key << "Script";
-            out << YAML::BeginMap;
-            out << MapEntry("Script", script.script);
-            out << MapEntry("Active", script.active);
+            out << YAML::Key << "Script" << YAML::BeginMap;
+            out << YAML::Key << "Script" << YAML::Value << script.script;
+            out << YAML::Key << "Active" << YAML::Value << script.active;
             out << YAML::EndMap;
         }
         if (entity.Has<PlayerComponent>()) {
             const auto& player = entity.Get<PlayerComponent>();
-            out << YAML::Key << "Player";
-            out << YAML::BeginMap;
-            out << MapEntry("MovingForwards", player.movingForwards);
-            out << MapEntry("MovingBackwards", player.movingBackwards);
-            out << MapEntry("MovingLeft", player.movingLeft);
-            out << MapEntry("MovingRight", player.movingRight);
-            out << MapEntry("Jumping", player.jumping);
-            out << MapEntry("Direction", player.direction);
-            out << MapEntry("Yaw", player.yaw);
-            out << MapEntry("Pitch", player.pitch);
+            out << YAML::Key << "Player" << YAML::BeginMap;
+            out << YAML::Key << "MovingForwards" << YAML::Value << player.movingForwards;
+            out << YAML::Key << "MovingBackwards" << YAML::Value << player.movingBackwards;
+            out << YAML::Key << "MovingLeft" << YAML::Value << player.movingLeft;
+            out << YAML::Key << "MovingRight" << YAML::Value << player.movingRight;
+            out << YAML::Key << "Jumping" << YAML::Value << player.jumping;
+            out << YAML::Key << "Direction" << YAML::Value << player.direction;
+            out << YAML::Key << "Yaw" << YAML::Value << player.yaw;
+            out << YAML::Key << "Pitch" << YAML::Value << player.pitch;
             out << YAML::EndMap;
         }
         if (entity.Has<CameraComponent>()) {
             const auto& camera = entity.Get<CameraComponent>();
-            out << YAML::Key << "Camera";
-            out << YAML::BeginMap;
-            out << MapEntry("Projection", camera.projection);
-            out << MapEntry("Pitch", camera.pitch);
+            out << YAML::Key << "Camera" << YAML::BeginMap;
+            out << YAML::Key << "Projection" << YAML::Value << camera.projection;
+            out << YAML::Key << "Pitch" << YAML::Value << camera.pitch;
             out << YAML::EndMap;
         }
         if (entity.Has<SelectComponent>()) {
@@ -287,18 +94,16 @@ void Serialiser::Serialise(const std::string& file)
         }
         if (entity.Has<PathComponent>()) {
             const auto& path = entity.Get<PathComponent>();
-            out << YAML::Key << "Path";
-            out << YAML::BeginMap;
-            out << MapEntry("Markers", path.markers);
-            out << MapEntry("Speed", path.speed);
+            out << YAML::Key << "Path" << YAML::BeginMap;
+            //out << YAML::Key << "Markers" << YAML::Value << path.markers; TODO - Fix this
+            out << YAML::Key << "Speed" << YAML::Value << path.speed;
             out << YAML::EndMap;
         }
         if (entity.Has<GridComponent>()) {
             const auto& grid = entity.Get<GridComponent>();
-            out << YAML::Key << "Grid";
-            out << YAML::BeginMap;
-            out << MapEntry("X", grid.x);
-            out << MapEntry("Z", grid.z);
+            out << YAML::Key << "Grid" << YAML::BeginMap;
+            out << YAML::Key << "X" << YAML::Value << grid.x;
+            out << YAML::Key << "Z" << YAML::Value << grid.z;
             out << YAML::EndMap;
         }
         out << YAML::EndMap;
